aura_ctx-0.3.3/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aura-ctx"
version = "0.3.6"
description = "Your portable AI context. Carry your identity across every AI tool."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nok (Enoch Afanwoubo)", email = "enochafanbo@outlook.com" },
]
keywords = ["ai", "context", "memory", "mcp", "portable", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
extract = [
    "openai>=1.0",
    "tiktoken>=0.5",
]
mcp = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
aura = "aura.cli:app"

[project.urls]
Homepage = "https://github.com/WozGeek/aura-ctx"
Documentation = "https://wozgeek.github.io/aura-ctx"
Repository = "https://github.com/WozGeek/aura-ctx"
Issues = "https://github.com/WozGeek/aura-ctx/issues"

[tool.hatch.build.targets.wheel]
packages = ["aura"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
